Up ViolatedConsDiagnostic as a top-level class in util package.
